styling changes and small corrections
