+++ app/globals.css
@@ -81,9 +81,9 @@ body {
 rdp {
   --rdp-cell-size: 40px; /* Size of the day cells. */
   --rdp-caption-font-size: 18px; /* Font size for the caption labels. */
-  --rdp-accent-color: #0000ff; /* Accent color for the background of selected days. */
+  --rdp-accent-color: #42bbff; /* Accent color for the background of selected days. */
   --rdp-background-color: #e7edff; /* Background color for the hovered/focused elements. */
-  --rdp-accent-color-dark: #3003e1; /* Accent color for the background of selected days (to use in dark-mode). */
+  --rdp-accent-color-dark: #42bbff; /* Accent color for the background of selected days (to use in dark-mode). */
   --rdp-background-color-dark: #180270; /* Background color for the hovered/focused elements (to use in dark-mode). */
   --rdp-outline: 2px solid var(--rdp-accent-color); /* Outline border for focused elements */
   --rdp-outline-selected: 3px solid var(--rdp-accent-color); /* Outline border for focused _and_ selected elements */
@@ -94,7 +94,6 @@ rdp {
 .rdp-day-selected:focus {
   background-color: #42bbff;
   color: var(--rdp-selected-color);
-  accent-color: #4f48e6;
 }
 .rdp-caption_label {
   color: white;
