+++ components/shared/profile/edit/UploadPhoto.tsx
@@ -3,14 +3,15 @@ import React, { useEffect } from "react";
 
 import { useForm } from "react-hook-form";
 import Image from "next/image";
-import { FaCloudUploadAlt } from "react-icons/fa";
+import { LuUploadCloud } from "react-icons/lu";
 
 import { Input } from "@/components/ui/input";
 import { uploadImage } from "@/lib/actions/cloudinary.actions";
 interface UploadPhotoProps {
   image?: string;
+  className?: string;
 }
-const UploadPhoto = ({ image }: UploadPhotoProps) => {
+const UploadPhoto = ({ image, className }: UploadPhotoProps) => {
   const { register, handleSubmit, getValues, watch } = useForm();
   const watchInput = watch("file");
   const inputRef = React.useRef<HTMLInputElement | null>(null);
@@ -31,13 +32,13 @@ const UploadPhoto = ({ image }: UploadPhotoProps) => {
   }, [getValues, watchInput]);
   return (
     <>
-      <div className="inline-flex flex-row px-[30px]">
+      <div className="inline-flex flex-row items-center gap-x-[14px] space-y-2 align-middle">
         <Image
           src={image!}
           alt="profile picture"
           width={90}
           height={90}
-          className="inset-inline-start mr-[8px]"
+          className="inset-inline-start "
         />
         <form onSubmit={handleSubmit(onSubmit)}>
           <Input
@@ -46,15 +47,16 @@ const UploadPhoto = ({ image }: UploadPhotoProps) => {
             {...rest}
             ref={(e) => {
               ref(e);
-              inputRef.current = e; // you can still assign to ref
+              inputRef.current = e;
             }}
           />
           <button
             type="submit"
-            className="paragraph-3-medium flex h-[40px] w-[200px] flex-row items-center justify-center gap-2 rounded-[5px] border-[1px] bg-black-700 px-3.5 py-2 text-white-300  shadow shadow-gray-800/10 "
+            className="inset-inline-end paragraph-3-medium r flex h-[40px] w-[200px] flex-row items-center gap-2 rounded-[5px] bg-black-700 px-3.5 py-2 align-middle text-white-300  shadow shadow-gray-800/10 "
           >
             {" "}
-            <FaCloudUploadAlt size={32} /> Upload a Picture
+            <LuUploadCloud size={32} />
+            <span className="text-nowrap">Upload Profile Picture</span>
           </button>
         </form>
       </div>
